requirements.txt
--find-links wheelhouse
Flask
requests
jsonlib-python3
azure-storage-blob
azure-cognitiveservices-vision-computervision
